# Conflicts:
#	include/isa_shunt.h
#	src/isa_shunt.cpp

+++ src/isa_shunt.cpp
@@ -24,7 +24,7 @@ int32_t ISA::KWh;
 int32_t ISA::Voltage=0;
 int32_t ISA::Voltage2=0;
 int32_t ISA::Voltage3=0;
-int16_t ISA::Temperature;
+int32_t ISA::Temperature;
 
 
 
@@ -260,25 +260,20 @@ void ISA::handle524(uint32_t data[2])  //Voltage3
 }
 
 void ISA::handle525(uint32_t data[2])  //Temperature
-
 {
    uint8_t* bytes = (uint8_t*)data;// arrgghhh this converts the two 32bit array into bytes. See comments are useful:)
    framecount++;
-   uint16_t temp=0;
-   temp = (uint16_t)((bytes[5] << 24) | (bytes[4] << 16) | (bytes[3] << 8) | (bytes[2]));
+   int32_t temp=0;
+   temp = (int32_t)((bytes[5] << 24) | (bytes[4] << 16) | (bytes[3] << 8) | (bytes[2]));
 
    Temperature=temp/10;
 
 }
 
-
-
 void ISA::handle526(uint32_t data[2]) //Kilowatts
-
 {
    uint8_t* bytes = (uint8_t*)data;// arrgghhh this converts the two 32bit array into bytes. See comments are useful:)
    KW = (int32_t)((bytes[5] << 24) | (bytes[4] << 16) | (bytes[3] << 8) | (bytes[2]));
-
 }
 
 
@@ -287,7 +282,6 @@ void ISA::handle527(uint32_t data[2]) //Ampere-Hours
 {
    uint8_t* bytes = (uint8_t*)data;// arrgghhh this converts the two 32bit array into bytes. See comments are useful:)
    Ah = (bytes[5] << 24) | (bytes[4] << 16) | (bytes[3] << 8) | (bytes[2]);
-
 }
 
 void ISA::handle528(uint32_t data[2])  //kiloWatt-hours
